appointment_app/requirement.txt
asgiref==3.6.0
Django==3.2.8
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.3